# openapi.yaml
openapi: 3.1.0
info:
  title: Smart Home Symcon Service
  description: API сервис управления устройствами умного дома.
  contact:
    name: Alexey Kuznetsov
    url: https://github.com/alexeycoder
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://localhost:8080/v1
    description: Test server
  - url: https://symcon.ru/v1
    description: Production server
...
paths:
...
  /groups:
    post:
      tags:
        - groups
      summary: Добавить группу
      description: Добавить группу пользователей совместного пользования устройствами
      operationId: addGroup
      parameters:
        - name: username
          in: header
          description: Имя для входа пользователя
          required: true
          schema:
            type: string
            format: email
      requestBody:
        description: Создание новой группы пользоватей в базе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '201':
          description: Created successfully
        '400':
          description: Invalid group data supplied
        '405':
          description: Invalid input
  /groups/{groupId}:
    get:
      tags:
        - groups
      summary: Найти группу по идентификатору
      description: вернуть группу
      operationId: findGroupById
      parameters:
        - name: username
          in: header
          description: Имя для входа пользователя
          required: true
          schema:
            type: string
            format: email
        - name: groupId
          in: path
          description: Идентификатор группы
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Group not found

    put:
      tags:
        - groups
      summary: Обновить группу
      description: Обновить группу пользователей совместного пользования устройствами
      operationId: updateGroup
      parameters:
        - name: username
          in: header
          description: Имя для входа пользователя
          required: true
          schema:
            type: string
            format: email
        - name: groupId
          in: path
          description: Идентификатор группы
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Обновить группу
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Group not found

    delete:
      tags:
        - groups
      summary: Удаляет пользователя
      description: Удаляет пользователя
      operationId: deleteGroup
      parameters:
        - name: username
          in: header
          description: Имя для входа пользователя
          required: true
          schema:
            type: string
            format: email
        - name: groupId
          in: path
          description: Идентификатор группы
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
...
  /schedules/{scheduleId}:
    get:
      tags:
        - schedules
      summary: Найти график работ по идентификатору
      description: Найти график работ по идентификатору
      operationId: findScheduleById
      parameters:
        - name: scheduleId
          in: path
          description: Идентификатор расписания
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Schedule not found

    put:
      tags:
        - schedules
      summary: Обновить график работ
      description: Обновить график работ
      operationId: updateSchedule
      parameters:
        - name: scheduleId
          in: path
          description: Идентификатор расписания
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Обновить график работ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Schedule not found
components:
  schemas:
...
    Group:
      required:
        - name
        - members
      type: object
      properties:
        groupId:
          type: integer
          format: int64
          examples:
            - 123
        name:
          type: string
          examples:
            - "Семья"
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: integer
                format: int64
          examples:
            - "- userId: 5
               - userId: 8
               "
...
    Schedule:
      required:
        - deviceId
        - dayOfWeek
        - timeOfDay
        - serviceTypeId
        - servicePropertiesCode
      type: object
      properties:
        scheduleId:
          type: integer
          format: int64
          examples:
            - 123
        dayOfWeek:
          type: integer
          format: int32
        timeOfDay:
          type: string
          format: time
          examples:
            - "10:35"
        serviceTypeId:
          type: integer
          format: int32
        servicePropertiesCode:
          type: integer
          format: int64
          examples:
            - 0xFFFFFFFF