# spring-server/src/main/java/io/swagger/api/SchedulesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.62).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Schedule;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2024-09-27T16:36:05.325149630Z[GMT]")
@Validated
public interface SchedulesApi {

    @Operation(summary = "Найти график работ по идентификатору", description = "Найти график работ по идентификатору", tags={ "schedules" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Schedule.class))),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
        
        @ApiResponse(responseCode = "401", description = "Unauthorized"),
        
        @ApiResponse(responseCode = "404", description = "Schedule not found") })
    @RequestMapping(value = "/schedules/{scheduleId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Schedule> findScheduleById(@Parameter(in = ParameterIn.PATH, description = "Идентификатор расписания", required=true, schema=@Schema()) @PathVariable("scheduleId") Object scheduleId
);


    @Operation(summary = "Обновить график работ", description = "Обновить график работ", tags={ "schedules" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK"),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
        
        @ApiResponse(responseCode = "401", description = "Unauthorized"),
        
        @ApiResponse(responseCode = "404", description = "Schedule not found") })
    @RequestMapping(value = "/schedules/{scheduleId}",
        consumes = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<Void> updateSchedule(@Parameter(in = ParameterIn.PATH, description = "Идентификатор расписания", required=true, schema=@Schema()) @PathVariable("scheduleId") Object scheduleId
, @Parameter(in = ParameterIn.DEFAULT, description = "Обновить график работ", required=true, schema=@Schema()) @Valid @RequestBody Schedule body
);

}